lectia 4

+++ app/Http/Controllers/BoardController.php
@@ -3,16 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Models\Board;
-use App\Models\User;
+use App\Models\BoardUser;
 use App\Models\Task;
+use App\Models\User;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\DB;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 /**
  * Class BoardController
@@ -30,7 +30,7 @@ class BoardController extends Controller
         $user = Auth::user();
 
         $boards = Board::with(['user', 'boardUsers']);
-        
+
         if ($user->role === User::ROLE_USER) {
             $boards = $boards->where(function ($query) use ($user) {
                 //Suntem in tabele de boards in continuare
@@ -47,53 +47,96 @@ class BoardController extends Controller
         return view(
             'boards.index',
             [
-                'boards' => $boards
+                'boards' => $boards,
+                'userList' => User::select(['id', 'name'])->get()->toArray()
             ]
         );
     }
 
-   
+    /**
+     * @param  Request  $request
+     * @param $id
+     *
+     * @return JsonResponse
+     */
     public function updateBoard(Request $request, $id): JsonResponse
     {
+        /** @var Board $board */
         $board = Board::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
 
         $error = '';
         $success = '';
 
         if ($board) {
-            $name = $request->get('name');
-            
-                $board->name = $name;
+            if ($board->user->id === $user->id || $user->role === User::ROLE_ADMIN) {
+                $newBoardUsers = $request->get('boardUsers');
+                $existingBoardUsers = [];
+
+                $board->boardUsers()->get()->each(function ($boardUser) use ($newBoardUsers, &$existingBoardUsers) {
+                    if (!in_array($boardUser->user_id, $newBoardUsers)) {
+                        $boardUser->delete();
+                    } else {
+                        $existingBoardUsers[] = $boardUser->user_id;
+                    }
+                });
+
+                $toSave = array_diff($newBoardUsers, $existingBoardUsers);
+
+                foreach ($toSave as $userId) {
+                    $boardUser = new BoardUser();
+                    $boardUser->board_id = $board->id;
+                    $boardUser->user_id = $userId;
+                    $boardUser->save();
+                }
+
+                $board->name = $request->get('name');
                 $board->save();
                 $board->refresh();
 
                 $success = 'Board saved';
-            
+            } else {
+                $error = 'You don\'t have permission to edit this board!';
+            }
         } else {
             $error = 'Board not found!';
         }
 
         return response()->json(['error' => $error, 'success' => $success, 'board' => $board]);
     }
-    
-        public function deleteBoard(Request $request, $id): JsonResponse
-        {
-            $board = Board::find($id);
-    
-            $error = '';
-            $success = '';
-    
-            if ($board) {
+
+    /**
+     * @param $id
+     *
+     * @return JsonResponse
+     */
+    public function deleteBoard($id): JsonResponse
+    {
+        /** @var Board $board */
+        $board = Board::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
+
+        $error = '';
+        $success = '';
+
+        if ($board) {
+            if ($board->user->id === $user->id || $user->role === User::ROLE_ADMIN) {
                 $board->delete();
-    
+
                 $success = 'Board deleted';
             } else {
-                $error = 'Board not found!';
+                $error = 'You don\'t have permission to delete this board!';
             }
-    
-            return response()->json(['error' => $error, 'success' => $success]);
-        }    
-   
+        } else {
+            $error = 'Board not found!';
+        }
+
+        return response()->json(['error' => $error, 'success' => $success]);
+    }
 
     /**
      * @param $id
@@ -106,7 +149,6 @@ class BoardController extends Controller
         $user = Auth::user();
 
         $boards = Board::query();
-        
 
         if ($user->role === User::ROLE_USER) {
             $boards = $boards->where(function ($query) use ($user) {
@@ -117,29 +159,18 @@ class BoardController extends Controller
             });
         }
 
-
         $board = clone $boards;
         $board = $board->where('id', $id)->first();
-        // dd($board->id); die();
+
         $boards = $boards->select('id', 'name')->get();
 
         if (!$board) {
             return redirect()->route('boards.all');
         }
 
-        // $tasks = Task::query();
+        $tasks = $board->tasks()->oldest()->paginate(10);
 
-        // $tasks = $tasks->where(function ($query) use ($board) {
-        //     $query->where('board_id',$board->id);
-        // });
-        
-        //dd($tasks); die();
-        //$tasks = DB::table('tasks')->with('user');
-        //$tasks = Task::with('user')->where('board_id',$id);
-        
-        $tasks = DB::select('select * from tasks where board_id = ?', [$board->id]);
-        
-        
+        $boardUsers = $board->boardUsers()->with('user')->get();
 
         return view(
             'boards.view',
@@ -147,7 +178,77 @@ class BoardController extends Controller
                 'board' => $board,
                 'boards' => $boards,
                 'tasks' => $tasks,
+                'boardUsers' => $boardUsers
             ]
         );
+    }
+
+    /**
+     * @param  Request  $request
+     * @param $id
+     *
+     * @return JsonResponse
+     */
+    public function updateTask(Request $request, $id): JsonResponse
+    {
+        /** @var Task $task */
+        $task = Task::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
+
+        $boardUser = BoardUser::where('board_id', $task->board_id)->where('user_id', $user->id)->first();
+
+        $error = '';
+        $success = '';
+
+        if ($task) {
+            if ($boardUser) {
+                $task->name = $request->get('name');
+                $task->description = $request->get('description');
+                $task->assignment = $request->get('assignment');
+                $task->status = $request->get('status');
+                $task->save();
+
+                $success = 'Task saved';
+            } else {
+                $error = 'You don\'t have permission to edit this task!';
+            }
+        } else {
+            $error = 'Task not found!';
+        }
+
+        return response()->json(['error' => $error, 'success' => $success, 'task' => $task]);
+    }
+
+    /**
+     * @param $id
+     *
+     * @return JsonResponse
+     */
+    public function deleteTask($id): JsonResponse
+    {
+        /** @var Task $task */
+        $task = Task::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
+
+        $error = '';
+        $success = '';
+
+        if ($task) {
+            if ($task->board->user->id === $user->id || $user->role === User::ROLE_ADMIN) {
+                $task->delete();
+
+                $success = 'Task deleted';
+            } else {
+                $error = 'You don\'t have permission to delete this task!';
+            }
+        } else {
+            $error = 'Task not found!';
+        }
+
+        return response()->json(['error' => $error, 'success' => $success]);
     }
 }
